chore: rename AutoStakeMaturityModal into NnsAutoStakeMaturityModal
