+++ frontend/src/lib/modals/neurons/NnsAutoStakeMaturityModal.svelte
